Conditionally dl binaries

+++ deploy.sh
@@ -1,16 +1,28 @@
 #!/bin/bash
 echo "User `whoami`"
-echo 'Install JQ'
-wget -O jq https://github.com/stedolan/jq/releases/download/jq-1.5/jq-linux64
-chmod u+x ./jq
+sonar_version='9.6.1.59531'
+sonar_home="`pwd`/sonarqube-${sonar_version}"
+
+if [ ! -f jq ]
+then
+	echo 'Install JQ'
+	wget -O jq https://github.com/stedolan/jq/releases/download/jq-1.5/jq-linux64
+	chmod u+x ./jq
+fi
+
+if [ ! -d $sonar_home ]
+then
+	echo "Install ${sonar_home}"
+	wget --no-hsts "https://binaries.sonarsource.com/Distribution/sonarqube/sonarqube-${sonar_version}.zip"
+	unzip "sonarqube-${sonar_version}.zip"
+fi
 
 # echo "Copy $DEPLOYMENT_SOURCE/wwwroot folder to $DEPLOYMENT_TARGET" 
 # cp -r $DEPLOYMENT_SOURCE/wwwroot/** $DEPLOYMENT_TARGET
 
-sonar_version='9.6.1.59531'
-wget --no-hsts "https://binaries.sonarsource.com/Distribution/sonarqube/sonarqube-${sonar_version}.zip"
-unzip -n "sonarqube-${sonar_version}.zip"
-export SONARQUBE_HOME="`pwd`/sonarqube-${sonar_version}"
+
+
+export SONARQUBE_HOME=$sonar_home
 export JAVA_HOME=/opt/java/openjdk
 export SONAR_JAVA_PATH="${JAVA_HOME}/bin/java"
